@@ -4,14 +4,14 @@
 	"private": true,
 	"dependencies": {
 		"@testing-library/jest-dom": "^5.16.4",
-		"@testing-library/react": "^10.4.9",
-		"@testing-library/user-event": "^10.4.1",
-		"axios": "^0.21.4",
-		"js-cookie": "^2.2.1",
-		"react": "^16.14.0",
-		"react-dom": "^16.14.0",
-		"react-router-dom": "^5.3.3",
-		"react-scripts": "^3.4.4",
+		"@testing-library/user-event": "^14.2.0",
+		"axios": "^0.27.2",
+		"js-cookie": "^3.0.1",
+		"react": "^18.1.0",
+		"react-dom": "^18.1.0",
+		"react-redux": "^8.0.2",
+		"react-router-dom": "^6.3.0",
+		"react-scripts": "^2.1.3",
 		"validator": "^13.7.0"
 	},
 	"scripts": {
@@ -37,5 +37,8 @@
 			"not op_mini all"
 		]
 	},
-	"proxy": "http://localhost:5000"
+	"proxy": "http://localhost:5000",
+	"devDependencies": {
+		"@testing-library/react": "^12.1.2"
+	}
 }
