@@ -1,5 +1,5 @@
 import React from 'react';
-import { BrowserRouter, Switch, Route } from 'react-router-dom';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import './App.css';
 import Header from './Header';
 import Home from './Home';
@@ -15,41 +15,49 @@ import AdminEditProduct from './AdminEditProduct';
 import AdminRoute from './AdminRoute';
 import UserRoute from './UserRoute';
 import NotFound from './NotFound';
+// import { isAuthenticated } from '../helpers/auth';
 
 const App = () => {
 	return (
 		<BrowserRouter>
 			<Header />
 			<main>
-				<Switch>
-					<Route exact path='/' component={Home} />
-					<Route exact path='/shop' component={Shop} />
-					<Route exact path='/cart' component={Cart} />
+				<Routes>
+					<Route exact path='/' element={<Home />} />
+					<Route exact path='/shop' element={<Shop />} />
+					<Route exact path='/cart' element={<Cart />} />
 					<Route
 						exact
 						path='/product/:productId'
-						component={Product}
+						element={<Product />}
 					/>
-					<Route exact path='/shipping' component={Shipping} />
-					<Route exact path='/signup' component={Signup} />
-					<Route exact path='/signin' component={Signin} />
-					<UserRoute
-						exact
-						path='/user/dashboard'
-						component={UserDashboard}
-					/>
-					<AdminRoute
-						exact
-						path='/admin/dashboard'
-						component={AdminDashboard}
-					/>
-					<AdminRoute
-						exact
-						path='/admin/edit/product/:productId'
-						component={AdminEditProduct}
-					/>
-					<Route component={NotFound} />
-				</Switch>
+					<Route exact path='/shipping' element={<Shipping />} />
+					<Route exact path='/signup' element={<Signup />} />
+					<Route exact path='/signin' element={<Signin />} />
+
+					<Route element={<UserRoute />}>
+						<Route
+							exact
+							path='/user/dashboard'
+							element={<UserDashboard />}
+						/>
+					</Route>
+
+					<Route element={<AdminRoute />}>
+						<Route
+							exact
+							path='/admin/dashboard'
+							element={<AdminDashboard />}
+						/>
+						<Route
+							exact
+							path='/admin/edit/product/:productId'
+							element={<AdminEditProduct />}
+						/>
+					</Route>
+
+					<Route element={<NotFound />} />
+				</Routes>
 			</main>
 		</BrowserRouter>
 	);
