@@ -1,10 +1,13 @@
 import React, { useEffect } from 'react';
+import { useNavigate, useParams } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import { getProduct } from '../redux/actions/productActions';
 import { addToCart } from '../redux/actions/cartActions';
 
-const Product = ({ match, history }) => {
-	const { productId } = match.params;
+const Product = () => {
+	let navigate = useNavigate();
+	let { productId } = useParams();
+
 	const dispatch = useDispatch();
 
 	useEffect(() => {
@@ -18,7 +21,7 @@ const Product = ({ match, history }) => {
 	};
 
 	const handleGoBackBtn = () => {
-		history.goBack();
+		navigate(-1);
 	};
 
 	return (
